:ambulance: Correction des conversions de l'ELO

+++ src/App/EloCalculator.php
@@ -9,11 +9,11 @@ class EloCalculator
     /**
      * @brief Calcule le nouvel Elo du joueur en fonction de son Elo actuel, de l'Elo moyen des joueurs de la partie et du résultat de la partie
      * @param $eloPlayer int Elo du joueur
-     * @param $averageElo int Elo moyen des joueurs de la partie
+     * @param $averageElo float Elo moyen des joueurs de la partie
      * @param $result float Résultat de la partie (1 pour une victoire, 0.5 pour un match nul, 0 pour une défaite)
      * @return float Nouvel Elo du joueur
      */
-    public static function calculateNewElo(int $eloPlayer, int $averageElo, float $result): float
+    public static function calculateNewElo(int $eloPlayer, float $averageElo, float $result): float
     {
         return $eloPlayer + self::K_FACTOR * ($result - self::getExpectedScore($eloPlayer, $averageElo));
     }
@@ -21,10 +21,10 @@ class EloCalculator
     /**
      * @brief Calcule la probabilité de victoire en fonction de l'Elo passé en paramètre
      * @param $eloPlayer int Elo du joueur
-     * @param $averageElo int Elo moyen des joueurs de la partie
+     * @param $averageElo float Elo moyen des joueurs de la partie
      * @return float Probabilité de victoire
      */
-    protected static function getExpectedScore(int $eloPlayer, int $averageElo): float
+    protected static function getExpectedScore(int $eloPlayer, float $averageElo): float
     {
         return 1 / (1 + pow(10, ($averageElo - $eloPlayer) / 400));
     }
